[refactor] 建立 PenCards 元件資料夾集中管理 PenCards 子元件
