+++ src/views/Trending.vue
@@ -73,7 +73,7 @@ import { Navigation } from "swiper/modules";
 import "swiper/css";
 import "swiper/css/navigation";
 import api from "@/config/api";
-import PenCard from "@/components/PenCard.vue";
+import PenCard from "@/components/PenCards/PenCard.vue";
 
 const swiperRef = ref(null);
 const pages = ref([]);
